Fixed issue causing break in program when querying mdx when creating table of content
This was due to the program looping through the nested array of headers using recursion without testing if there was a header present to style in the table of content when/before to comparing it to the current intersecting id of a header. This would break e.g when there was a h3 tag but no h2 tag above it in the nested array of id headers as it would try to slice an undefined value instead of the usual string.